+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 All notable changes to voice-input.
-## [unreleased]
+## v1.0.50 — 2026-05-04
 ### 2026-05-04
+- ![chore](https://img.shields.io/badge/chore-cfd3d7?style=flat-square) [`a6db0a8`](https://github.com/artur-arc/voice-input/commit/a6db0a8ccb4f185ef52d787f132577d53b2676d5) Release v1.0.50
 - ![feat](https://img.shields.io/badge/feat-0075ca?style=flat-square) [`4056a33`](https://github.com/artur-arc/voice-input/commit/4056a33a97fe69567239f17b71ddfca9d6f3b9a8) Add model information menu item
 - ![refactor](https://img.shields.io/badge/refactor-e4e669?style=flat-square) [`ce437ac`](https://github.com/artur-arc/voice-input/commit/ce437ac914335fe9b0e181cf6e5212c124352c6e) Update model selection and download logic
 - ![refactor](https://img.shields.io/badge/refactor-e4e669?style=flat-square) [`2c2569a`](https://github.com/artur-arc/voice-input/commit/2c2569a395acca8ef5a56bb1921ebdac8264d982) Group sound and notifications under